fix: 경로 수정 (`/result → /result?result={id}`)
