+++ pages/result/[resultId].tsx
@@ -1,89 +1,90 @@
-import { Button, MessageBox, Title } from '@components/common';
-import { Card } from '@components/result';
+import { Button, Description, MessageBox, Title } from '@components/common';
+import { shareKakaoLink, shareOtherLinks } from '@components/result';
 import styled from '@emotion/styled';
-import Head from 'next/head';
-import Router from 'next/router';
+import { DuckDomWithImg } from '@shared/types/DuckDam';
+import { theme } from '@styles/index';
+import Image from 'next/image';
+import { useRouter } from 'next/router';
+import loadingGif from 'public/images/loading-done.gif';
+import { useEffect, useState } from 'react';
 
-type ResultProps = {
-    data: DuckdamType;
-};
-type DuckdamType = {
-    img_url: string;
-    first_word: string;
-    second_word: string;
-    third_word: string;
-};
+const LodingImageSize = 188;
+const Result = () => {
+    const router = useRouter();
+    const [duckdam, setDuckdam] = useState<DuckDomWithImg>();
 
-const ResultDetail = ({ data }: ResultProps) => {
-    const { img_url, first_word, second_word, third_word } = data;
+    useEffect(() => {
+        const dataFetch = async () => {
+            const { result } = router.query;
+            const res = await fetch(`/api/duckdam/${result}`);
+            const data = await res.json();
+            setDuckdam(data);
+        };
+
+        dataFetch();
+    }, []);
 
     return (
         <>
-            <Head>
-                <title>사실은...</title>
-                <meta
-                    name="description"
-                    content="토끼가 만들어준 덕담을 확인해보세요! 장난꾸러기 토끼와 재밌고 즐거운 한가위 보내시길 바랍니다🐰"
-                />
-                <meta name="keyword" content="즐거운, 토끼" />
-                <meta property="og:image" content={img_url} />
-                <meta property="og:image:width" content="1200" />
-                <meta property="og:image:height" content="630" />
-                <meta property="og:title" content="비밀 덕담이 도착했습니다!" />
-                <meta
-                    property="og:description"
-                    content="여기를 눌러 덕담을 확인하세요 💌"
-                />
-                <meta name="twitter:card" content="summary" />
-                <meta
-                    name="twitter:title"
-                    content="비밀 덕담이 도착했습니다!"
-                />
-                <meta
-                    name="twitter:description"
-                    content="여기를 눌러 덕담을 확인하세요 💌"
-                />
-                <meta name="twitter:image" content={img_url} />
-            </Head>
             <Wrapper>
-                <Title>내 진짜 마음은...</Title>
+                <Image
+                    src={loadingGif}
+                    width={LodingImageSize}
+                    height={LodingImageSize}
+                    alt="로딩 완료 이미지"
+                />
+
                 <MessageBox>
-                    <Card>{first_word}</Card>
-                    <Card>{second_word}</Card>
-                    <Card>{third_word}</Card>
+                    <Title>덕담이 만들어졌어요!</Title>
+                    <Description
+                        text={
+                            '*남들에게 공유하기 전까지 덕담은 비밀로 유지됩니다!'
+                        }
+                    ></Description>
                 </MessageBox>
+                <Button>링크 복사하기</Button>
+            </Wrapper>
+            <ShareWrapper>
                 <Button
                     onClick={() => {
-                        Router.push('/');
+                        if (duckdam) {
+                            const props = {
+                                imageURL: duckdam.img_url,
+                            };
+                            shareKakaoLink(props);
+                        }
                     }}
+                    backgroundColor={theme.color.orange}
                 >
-                    나도 덕담 나눌래!
+                    카톡 공유하기
                 </Button>
-            </Wrapper>
+                <Button
+                    onClick={() => {
+                        shareOtherLinks();
+                    }}
+                    backgroundColor={theme.color.orange}
+                >
+                    다른 앱으로 공유하기
+                </Button>
+            </ShareWrapper>
         </>
     );
 };
 
-export default ResultDetail;
-
-export async function getServerSideProps(context: {
-    params: { resultId: string };
-}) {
-    const { params } = context;
-
-    const res = await fetch(
-        `${process.env.NEXT_PUBLIC_SITE_URL}/api/duckdam/${params.resultId}`
-    );
-    const data = await res.json();
-
-    return {
-        props: { data }, // will be passed to the page component as props
-    };
-}
+export default Result;
 
 const Wrapper = styled.div`
     width: 100%;
-    height: 100%;
+    height: 80%;
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+`;
+
+const ShareWrapper = styled.div`
+    position: relative;
+    width: 60%;
     display: flex;
     flex-direction: column;
     justify-content: center;
